+++ src/components/Header/Header.tsx
@@ -1,11 +1,11 @@
-import ProgressBar from '../atoms/ProgressBar';
-import NavBar from '../molecules/NavBar';
+import HeaderNav from './HeaderNav';
+import HeaderProgress from './HeaderProgress';
 
 function Header() {
   return (
     <div className="fixed left-0 top-0 flex w-screen max-tablet:h-screen max-tablet:w-auto tablet:flex-col">
-      <ProgressBar />
-      <NavBar />
+      <HeaderProgress />
+      <HeaderNav />
     </div>
   );
 }
