+++ src/components/Sections/ContactSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ContactSection';
