+++ src/components/Sections/IntroSection/IntroText.tsx
