+++ src/components/Sections/IntroSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './IntroSection';
