+++ src/components/Sections/ProjectsSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProjectsSection';
