+++ src/components/Sections/SkillsSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SkillsSection';
