+++ src/utils/util.ts
@@ -1,0 +1,2 @@
+export const capitalize = (str: string) =>
+  str[0].toUpperCase() + str.slice(1, str.length);
